fix synthesepy to accept newer installed version of CMake.
A check was done that CMake was installed but only a
strict version 2.8.5 was accepted. When not in this case,
this version was downloaded and compiled even if the system
had a newer version.

With this patch, if a newer version is installed we do nothing,
and the installed version is used.

git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8110 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de