refs #11347 basic working version
This version of the service (no service insert / delete and it does
not manage the timestamping of updates).

The service accept an unlimited number of changes by specifying the arguments
seN, atN and dtN with N being a number in sequence starting at 1.

Now the service use the stop area instead of the stop index in the service.
Issue : #11347


git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8155 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de