Add the error description in the http reply.
When we enter an invalid http service request,
for instance with a missing parameter there is a log
message with the missing parameter in the console but
not in the http reply. This patch add the description
in the http reply.

This patch also build the reply instead of having a
constant for each message.

git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8181 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de