added NCE retrieval of next departure time at terminus.
Now we catch the information messages from the NCE and
parse them to extract the next departure time.

Now in vehicle_informations, if the vehicle has the TERMINUS_START
status and that the NCE gave us the information, then a new parameter
is returned "terminus_departure_time" that contains a date.

bug refs #26714.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8286 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de