bug refs #27002 added 'relaxTransparency': 'maybe' option to polipo proxy.
In our case, if the polipo cache is on a persistant directory across reboot,
this will let the kiosk to be able to start even if SYNTHESE is either down
or not reachable.

Based on polipo user manual, this option does the following:

If relaxTransparency is maybe, all stale instances are still
validated, but a failure to connect is only reported as an error
if no data is available in the cache. If a connection fails and
stale data is available, it is served to the client with a
suitable HTTP/1.1 ‘Warning’ header. Current user-agents do not
provide visible indication of such warnings, however, and this
setting will typically cause the browser to display stale data
with no indication that anything went wrong. It is useful when
you are consulting a live web site but don't want to be bothered
with failed revalidations.


git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8342 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de