Added support for HTTP Caching by providing a MaxAge field on CMS Webpages.
By default nothing is changed on the SYNTHESE reply HTTP header
unless the max age field on a CMS web page is set.

When set, this will add the header:
Cache-Control: public, max-age=X

This will allow proxies and browsers to keep a version of this
page for max age duration. This can be used for example to avoid
having to serve on each resquest static content like images and
javascript.