Implements bug revs #27205 stop button file pooler
This new feature is especially usefull when a given hardware reports
that a passengers clicked on the stop button through a file in /proc.

To use this feature, just pass 2 parameters on the command line:
- stop_monitoring_file: the file to monitor like /proc/input-0
- stop_monitoring_period_ms: the polling period in milliseconds

When a 1 is read at the first line of this file, a call to 
VehicleModule::GetCurrentJourney().setStopRequested(true) is done, else
VehicleModule::GetCurrentJourney().setStopRequested(false) is called.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8363 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de