Improvements for the OSM importer. Taking into account more tags especially related to user class accessibility (access, foot, bicycle, motor_vehicles...) leading to cleverer road journey planning in general.
Improving greatly definition of a central road chunk for a city. Update at each OSM import and don't search to create one if there is a main stop area in the city. As for physical stop projection, ensure that the road chunk chosen is accessible to every user class.

Please reimport OSM datas after this commit to benefit from all the improvements.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8381 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de