fixes getNextService with routes having schedules after 24h
In Edge.cpp we keep a services running at this edge in a by hour slot
array in order to speed the processing. What happens in case of an
after midnight schedule, it is saved in the 0 slot. But it is also
the last schedule and it makes the main loop of getNextService
to end prematuraly.

This fix add a new case, on first run we try again with the next
slot to catch this situation.

fixed refs #26301



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8450 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de