Properly catch signals so that core dumps are generated.
In order to track crashes it is useful to configure the system
to generate core dumps. But it was not working properly because
signals where catched in SYNTHESE itself.

This patch re-raise the signals with our handler removed so
that the system properly catch it and generate core dumps.

This is documented here:
http://www.gnu.org/software/libc/manual/html_mono/libc.html#Termination-in-Handler