fixed a frequent crash case while adding/removing services.
In several places we loop over the services. But we also
can create or remove services, either from the admin site
or through the real time system.

There was not mutex to control this situation and the iteration
over the service list crashes when the list changes.

In this patch, we use the shared_recursive_mutex (including
in our code case). It provides a way to have a single writer lock
and a many reader lock. This ensure we won't lock too much in the
usual case where we compute several routing.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8523 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de