Major rework of the Session management.
In the previous implementation, we had a regular pointer to a
Session stored in a map. These Session instances are shareable
between threads. Uppon Session expiration we try to delete the
Session but it crashes if other threads have a lock on it
(the lock is an instance variable of Session).

In the new implementation, the map keeps shared_ptr of Session.
We no more explicitely delete Session, it will be deleted once
no thread references it and thus solves the crash case.