kiosk, added support to manually force a browser refresh.
Now we check that the online kiosk_config.json has changed.
If so, we refresh the browser.

At good practise is to add a line in the json file like:
"revision": 1,
and to increase the number each time you want to refresh the
browser.

This way you can still test a site by manually refreshing
your bowser and when you are happy with it modify the revision
number to force a browser refresh.

refs #27002


git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8599 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de