fixed a bug in InterSYNTHESE in case of a bad geometry
In case of a non valid geometry we where passing the string
POINT(inf inf) to the slave which creates a crash in spatialite.

This happens in case of a failure in the geometry convertion but
could happens in other conditions.

The fix prints a warning in the master logs and just transfer an empty
geometry to the slave.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8655 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de