Replaced the global_variable_update action by a service
The service has the same name, take the same parameters,
does to same thing except that it can also be used to
return the list of parameters and their current values.

The action 'global_variable_update' is marked as
deprecated.

This can be used to create an admin page that allows
the user to display and change these variables.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8725 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de