fixed site svn export
For some reasons it happens that site page may have the same rank.
In this case these pages was lost in the svn commit process.

this is due to our TreeNode storing items in a map thus hidding
items at the same tree rank.

Since the TreeNode is used in other places, there is a risk of
regression in DisplayScreen and CalendarTemplate.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/branches/3.8.beta1@8865 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de