Revision r8942 merged from branch 3.8.beta1
InterSYNTHESE, no more report each queue items in the admin

In InterSYNTHESE admin we have a view for the queue items.
Since it may be huge and that we display only the first one,
it is useless to return all the items.

Added a parameter 'queue_size' to let the user interface display
it.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@8943 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de