Merged Revision r8962 from 3.8.beta1
InterSYNTHESE, mark slave as updated as soon as the full dump is done.

If we don't do this and the slave crash during before the end of the initial
dump we had to redo the same long processing again. Since the data in is the
slave queue in memory and on disk the is no risk to to this.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9029 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de