Merged Revision r8985 from 3.8.beta1
InterSYNTHESE, process full update in background

When a long operation is performed in an HTTP worker thread,
SYNTHESE no more respond in a timely manners to client requests.

This maybe a bug or a 'feature' of asio. Maybe it is fixed in
a later release (currently tested on boost 1.42).

This patch runs the long operation 'full slave update' in a
specific thread and returns a nice message to the slave
inviting him to come back pick the results.

this does not break or change the overall communication
between the master and the slave. Just that now a full
update no more block the slave for minutes.

bug refs #30316



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9031 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de