Merged Revision r9039 from 3.8.beta1
Site svn connection, added options to make it useable.

This patch gives more power to the user to use the svn connector
to make it useable in many scenario where it is currently not
working.

SVN UPDATE:

Added a new option to let the user erase his own changes in the
site and with the ones from svn.

SVN COMMIT:

Added an option to let the user save his site to the working
copy without making a commit. This give the user a chance to
review what would be commited, fix it and commit manually.

Added an option to not do an update the working copy. This
is especially usefull with the previous option.