Bug fix on minMaxDurationRatioFilter.
Before this commit, if we made a RoutePlanner request with
5 results wanted (msn parameter) and a minMaxDurationRatio of 3 (min_max_duration_ratio_filter parameter),
PTTimeSlotRoutePlanner was going to return 5 results.
Then, we would apply a filter on duration ratio.
If some results were filtered, RoutePlannerFunction would return less than 5 results to the user.
The filtering is now done in TimeSlotRoutePlanner directly in order to have the right amount of results.