Added optional compilation of useless subdirectories.
By default the directories packages, projects, tools and utils
are no more processed by default by make. If you want them
you must configure manually select them at cmake time.

Added the option --do-not-build-python to synthesepy to
no more create and include a python environment and the
kiosk in the binary build. To be backward compatible,
this option is not activated by default.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9133 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de