Fixed IneoRT, now the path is properly saved.
There was a bug in which the path was not loaded in
FIELDS_ONLY_LOAD_LEVEL mode and thus was saved as 0.

Added a hack by using the linkLevel UNKNOWN_LOAD_LEVEL
in JourneyPatternTableSync to let it don't load
anythink. One can think of it as create the object with
its key only.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/branches/3.8.beta1@9175 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de