Feature : possible to use CMS page for external junctions in RoutePlanner
It is also a bugfix because if an external junction was in the result of rp using CMS page to display result, it was not possible to display the result.
There is no change by default (the ROAD page parameter is still called "junction_page").
The new parameters to use CMS page for external junctions are "external_junction_page" and "map_external_junction_page".

Work done by Xavier Raffin (Tisseo)

git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9188 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de