Feature : New PermanentThread / Device factory
This new feature allows to define devices to communicate with SYNTHESE.
It uses the new t115_permanent_threads table which registers for the devices SYNTHESE has to communicate (for example a validator).
When SYNTHESE is starting, the launchPoller() virtual method of the devices defined in t115_permanent_tables is launched.
Devices are classes inheriting DeviceTemplate class.
Issue : #32759


git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9332 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de