Feature : add new parameter "gleis_has_one_stop_per_line" for Hafas file format and some "catch" around lexical cast.
This parameter, if set to true, changes the way to read the gleis file.

When set to false, format is supposed to be
@Z 00002 000011 00
8503000 0 14     000000
8503206 0 2      000000
8509411 0 2      000000
8503209 0 3      000000
8509004 0 3      000000
8509002 0 4      000000
8503202 0 6      000000
8503225 0 6      000000
8509000 1 8      000000
8500010 2 9      000000
where the first line identifies the service and the following lines describe the stops.

When set to true, format is supposed to be like 
8509000 00002 000011 8        0916 057295
8509002 00002 000011 4             057295
8509004 00002 000011 3             057295
8509411 00002 000011 2             057295
where each line identifies the service and a stop
Issue : #34058


git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9478 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de