Feature : Max approach distance can now be set with multiple values in journey_planner service
"mad" param can now be written with multiple values in a call to journey_planner service.
For instance "&mad=100,1000,5000" will make a run with a max approach distance of 100 meters.
If no solution, a second run with a 1000 meters max approach distance will be done.
If there is still no solution a last run with a 5000 meters max approach distance will be done.

Passing 0 or 1 value to mad does not change the behavior of the service.
Issue : #34196


git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9495 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de