Feature : default sorting on Commercial Lines.
In the admin panel and in services output, all the commercial lines are now sorted the same way.
Rather than a random sort depending on multiple attributes a comparator is defined
in CommercialLine Class based on weight and short_name.

The weight is a new column in t042_commercial_lines (weight_for_sorting).
A new paramter is also available at startup "sort_letters_before_numbers" in order to
choose between two alphanumeric sorting : 1/3/10/A/C or A/C/1/3/10 (letters before/after numbers).

Commercial lines will be sorted by descending weight.
If 2 lines have the same weight they will be sorted by their short names
in an alphanumeric order (text first only if sort_lettres_before_numbers is set to 1)
except if the short_name is empty then we sort on the id.



git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9665 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de