Feature : add "complete_empty_stop_area_name" parameter.
If not set it does not change anything.
If set the stop areas with no names are registered as "name of the city + value of complete_empty_stop_area_name".
For instance if the KOORD file contains the line
0000050    690.937    192.315 447    % Altdorf
and the param is set to " (CFF)"
the new stop area will be named "Altdorf (CFF)" instead of no name.

git-svn-id: https://extranet.rcsmobility.com/svn/synthese3/trunk@9674 ae2791e7-e8fc-0310-8f54-ad5a55fdf1de